Add new view to router
